requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
langgraph
duckduckgo-search
python-dotenv
tavily-python
